Major update LibraryApp beta8 libs, viewcomponents, ...

+++ LibraryApp/App.Models/Ahs/JsonDetailSearch.cs
@@ -8,8 +8,8 @@ namespace App.Models.Ahs
     public class JsonDetailSearch
     {
         public string act { get; set; }
-        public Int32 totaal { get; set; }
-        public Int16 offset { get; set; }
+        public int totaal { get; set; }
+        public int offset { get; set; }
         public LibraryItem res { get; set; }
     }
 }
