+++ LibraryApp/App.Models/Ahs/JsonSimpleSearch.cs
@@ -8,8 +8,8 @@ namespace App.Models.Ahs
     public class JsonSimpleSearch
     {
         public string act { get; set; }
-        public Int32 totaal { get; set; }
-        public Int16 offset { get; set; }
+        public int totaal { get; set; }
+        public int offset { get; set; }
         public ICollection<JsonLibraryItem> res { get; set; }
     }
 }
